-made available day long in products

+++ src/main/java/org/gamein/marketservergamein2022/core/sharedkernel/entity/Product.java
@@ -47,7 +47,7 @@ public class Product {
     private Integer price;
 
     @Column(name = "available_day")
-    private Integer availableDay;
+    private Long availableDay;
 
     @ManyToOne
     private ResearchSubject RAndD;
